# controllers/user_controller.go
package controllers

/* Generated by egg v0.0.1 */

import (
	"fmt"

	"github.com/adamkali/mindscape/cmd/configuration"
	handlers "github.com/adamkali/mindscape/models/handlers/user_handlers"
	"github.com/adamkali/mindscape/services"
	"github.com/labstack/echo/v4"
)

type UserController struct {
	Name             string
	Config           *configuration.Configuration
	AuthService      services.IAuthService
	UserService      services.IUserService
	MinioService     services.IMinioService
	RedisService     services.IRedisService
	ValidatorService *services.ValidatorService
}

func (uc UserController) ControllerName() string {
	return uc.Name
}

func BuildUserController(p *Registrar) UserController {
	return UserController{
		Name:             "/users",
		Config:           p.Config,
		AuthService:      p.AuthService,
		MinioService:     p.MinioService,
		UserService:      p.UserService,
		RedisService:     p.RedisService,
		ValidatorService: p.ValidatorService,
	}
}

// @Summary Delete User by their UUID
// @Description get string by ID
//
// @ID          DeleteUserByUUID
// @Tags        Users
// @Produce     json
// @Param       user_id             path         string                         true "User Id"          default("e38e78a4-2ca3-4c59-a3ea-a2019866e593")
// @Param       Authorization       header       string                         true "admin header"     default("Bearer token")
// @Success     200                 {object}     responses.DeleteUserResponse
// @Router      /users/{user_id}    [delete]
func (UserController *UserController) DeleteUser(ctx echo.Context) error {
	return handlers.NewDeleteUserHandler(
		ctx,
		*UserController.ValidatorService,
		UserController.UserService,
		UserController.AuthService,
	).Handle().JSON()
}

// @Summary Signup to the app
// @Description Signup using the requests.NewUserRequest
//
// @ID          Signup
// @Tags        Users
// @Accept      json
// @Produce     json
// @Param       SignupRequest   body        NewUserRequest          true "Signup Request"
// @Success     200             {object}    responses.LoginResponse
// @Failure     400             {object}    responses.LoginResponse
// @Failure     500             {object}    responses.LoginResponse
// @Router      /users/signup   [post]
func (UserController *UserController) Signup(ctx echo.Context) error {
	return handlers.NewRegisterHandler(
		ctx,
		*UserController.ValidatorService,
		UserController.UserService,
		UserController.AuthService,
	).Handle().JSON()
}

// @Summary Login
// @Description to a user account with either email or username
//
// @ID          Login
// @Tags        Users
// @Accept      json
// @Produce     json
// @Param       LoginRequest body           LoginRequest            true "Log in request"
// @Success     200             {object}    responses.LoginResponse
// @Failure     400             {object}    responses.LoginResponse
// @Failure     401             {object}    responses.LoginResponse
// @Failure     500             {object}    responses.LoginResponse
// @Router      /users/login    [post]
func (uc *UserController) Login(ctx echo.Context) error {
	return handlers.NewLoginHandler(
		ctx,
		*uc.ValidatorService,
		uc.UserService,
		uc.AuthService,
	).Handle().JSON()
}

// @Summary Get Current User
// @Description Get the Current User by the uuid storred in the Claims header
//
// @ID          GetCurrentLoggedInUser
// @Tags        Users
// @Produce     json
// @Param       authorization   header       string                         true "admin header"     default(Bearer token)
// @Success     200             {object}     responses.UserResponse
// @Failure     400             {object}     responses.UserResponse
// @Failure     401             {object}     responses.UserResponse
// @Router      /users/current  [get]
func (UserController *UserController) GetCurrent(ctx echo.Context) error {
	return handlers.NewGetCurrentLoggedInUserHandler(
		ctx,
		*UserController.ValidatorService,
		UserController.UserService,
		UserController.AuthService,
	).Handle().JSON()
}

// @Summary		Upload file
// @Description	Upload file
//
// @ID				UploadProfilePicture
// @Tags            Users
// @Accept			multipart/form-data
// @Produce			json
// @Param			file				formData	file			true	"this is a test file"
// @Param           authorization		header      string          true	"admin header"        default(Bearer token)
// @Success		    200					{string}	UserResponse
// @Failure		    400					{object}	UserResponse
// @Failure		    404					{object}	UserResponse
// @Failure		    404					{object}	UserResponse
// @Router			/users/profile		[post]
func (UserController *UserController) UploadProfilePicture(ctx echo.Context) error {
	return handlers.NewUploadProfilePictureHandler(
		ctx,
		*UserController.ValidatorService,
		UserController.UserService,
		UserController.AuthService,
		UserController.MinioService,
	).Handle().JSON()
}

// @Summary Get User Profile by Authorization Header
// @Description Get User Profile by Authorization Header
//
// @ID          GetProfilePicture
// @Tags        Users
// @Produce     json
// @Param       Authorization       header       string                         true "admin header"     default(Bearer token)
// @Success     200                 {object}     responses.StringResponse
// @Failure     401                 {object}     responses.StringResponse
// @Failure     403                 {object}     responses.StringResponse
// @Failure     500                 {object}     responses.StringResponse
// @Router      /users/profile		[get]
func (uc *UserController) GetProfile(ctx echo.Context) error {
	return handlers.NewGetProfilPictureHandler(
		ctx,
		*uc.ValidatorService,
		uc.RedisService,
		uc.MinioService,
		uc.AuthService,
		uc.UserService,
	).Handle().JSON()
}

// @Summary Get All Users
// @Description Get All Users. Must be Admin using the new mediator pattern
//
// @ID          GetUsers
// @Tags        Users
// @Produce     json
// @Param       Authorization       header       string                         true "admin header"     default(Bearer token)
// @Success     200                 {object}     UsersResponse
// @Failure     403                 {object}     UsersResponse
// @Failure     404                 {object}     UsersResponse
// @Failure     500                 {object}     UsersResponse
// @Router      /users/			    [get]
func (uc *UserController) GetUsers(ctx echo.Context) error {
	return handlers.NewGetUsersHandler(
		ctx,
		uc.AuthService,
		uc.UserService,
	).Handle().JSON()
}

// @Summary Update User Credentials
// @Description Update User Credentials by id, Must be the same user in the jwt
// @Description that is being updated
//
// @ID          UpdateUser
// @Tags        Users
// @Accept      json
// @Produce     json
// @Param       UpdateCredentialsRequest body         UpdateCredentialsRequest       true "Update User Credentials Request"
// @Param       Authorization            header       string                         true "admin header"                    default(Bearer token)
// @Success     200                      {object}     UpdateUserResponse
// @Failure     400                      {object}     UpdateUserResponse
// @Failure     401                      {object}     UpdateUserResponse
// @Failure     500                      {object}     UpdateUserResponse
// @Router      /users/creds             [post]
func (uc *UserController) UpdateUser(ctx echo.Context) error {
	return handlers.NewUpdateUserHandler(
		ctx,
		*uc.ValidatorService,
		uc.UserService,
		uc.AuthService,
	).Handle().JSON()
}

// @Summary Get Default Background
// @Description Get Default Background
//
// @ID          GetDefaultBackground
// @Tags        Background
// @Produce     json
// @Success     200                 {object}     responses.StringResponse
// @Failure     404                 {object}     responses.StringResponse
// @Failure     500                 {object}     responses.StringResponse
// @Router      /background   [get]
func (uc *UserController) GetDefaultBackground(ctx echo.Context) error {
	return handlers.NewDefaultBackgroundHandler(
		ctx,
		*uc.ValidatorService,
		uc.RedisService,
		uc.MinioService,
	).Handle().JSON()
}

// @Summary Upload Background
// @Description Upload a Background and set the User.Background  to be the uploaded background
//
// @ID          UploadBackground
// @Tags        Users
// @Accept      multipart/form-data
// @Param       file                formData    file                            true "this is a test file"
// @Param       Authorization       header       string                         true "admin header"     default(Bearer token)
// @Produce     json
// @Success     200                 {object}     responses.StringResponse
// @Failure     404                 {object}     responses.StringResponse
// @Failure     500                 {object}     responses.StringResponse
// @Router      /users/background   [post]
func (uc UserController) UploadBackgound(ctx echo.Context) error {
	return handlers.NewUploadBackgroundHandler(
		ctx,
		*uc.ValidatorService,
		uc.UserService,
		uc.AuthService,
		uc.RedisService,
		uc.MinioService,
	).Handle().JSON()
}

// #region Background

// @Summary Get Background Choices that ae loaded into the server.
// @Description Get Background Choices
//
// @ID          GetBackgroundChoices
// @Tags        Background
// @Produce     json
// @Success     200                 {object}     responses.BackgroundsResponse
// @Failure     404                 {object}     responses.BackgroundsResponse
// @Failure     500                 {object}     responses.BackgroundsResponse
// @Router      /background/choices   [get]
func (uc *UserController) GetBackgroundChoices(ctx echo.Context) error {
	return handlers.NewBackgroundChoicesHandler(
		ctx,
		*uc.ValidatorService,
		uc.RedisService,
		uc.MinioService,
	).Handle().JSON()
}

// @Summary Get User Background Choice
// @Description Get User Background Choice
//
// @ID          GetUserBackground
// @Tags        Users
// @Param       Authorization       header       string                         true "admin header"     default(Bearer token)
// @Param       background          query        string                         true "background"
// @Produce     json
// @Success     200                 {object}     responses.StringResponse
// @Failure     404                 {object}     responses.StringResponse
// @Failure     500                 {object}     responses.StringResponse
// @Router      /users/background   [get]
func (uc UserController) GetBackground(ctx echo.Context) error {
	return handlers.NewBackgroundHandler(
		ctx,
		*uc.ValidatorService,
		uc.AuthService,
		uc.MinioService,
	).Handle().JSON()

}

// @Summary Get User Background Choices
// @Description Get User Backgrounds Uploaded to the server
//
// @ID          GetUserBackgroundChoices
// @Tags        Users
// @Param       Authorization       header       string                         true "admin header"     default(Bearer token)
// @Produce     json
// @Success     200                 {object}     responses.BackgroundsResponse
// @Failure     404                 {object}     responses.BackgroundsResponse
// @Failure     500                 {object}     responses.BackgroundsResponse
// @Router      /users/background/choices   [get]
func (uc UserController) GetUserBackgroundChoices(ctx echo.Context) error {
	fmt.Println("[INFO] GetUserBackgroundChoices")
	return handlers.NewUserBackgroundChoicesHandler(
		ctx,
		*uc.ValidatorService,
		uc.AuthService,
		uc.MinioService,
	).Handle().JSON()
}

// @Summary Set User Background
// @Description Set User Background Choice by background name
// @Description by query param.
//
// @ID          SetUserBackground
// @Tags        Users
// @Param       Authorization       header       string                         true "admin header"     default(Bearer token)
// @Param       background          query        string                         true "background"
// @Produce     json
// @Success     200                 {object}     responses.StringResponse
// @Failure     404                 {object}     responses.StringResponse
// @Failure     500                 {object}     responses.StringResponse
// @Router      /users/background   [patch]
func (uc UserController) SetBackground(ctx echo.Context) error {
	return handlers.NewSetUserBackgroundHandler(
		ctx,
		uc.UserService,
		uc.AuthService,
		uc.MinioService,
	).Handle().JSON()
}

func (uc UserController) Attatch(e *echo.Echo, authMiddleware echo.MiddlewareFunc) {
	e.GET("/api/background", uc.GetDefaultBackground)
	e.GET("/api/background/choices", uc.GetBackgroundChoices)

	// Register the namespaces for the endopoints
	api := e.Group("/api" + uc.Name)
	api.GET("", uc.GetUsers, authMiddleware)
	api.POST("/login", uc.Login)
	api.POST("/signup", uc.Signup)
	api.GET("/current", uc.GetCurrent, authMiddleware)
	api.POST("/profile", uc.UploadProfilePicture, authMiddleware)
	api.POST("/creds", uc.UpdateUser, authMiddleware)
	api.GET("/profile", uc.GetProfile, authMiddleware)
	api.POST("/background", uc.UploadBackgound, authMiddleware)
	api.PATCH("/background", uc.SetBackground, authMiddleware)
	api.GET("/background", uc.GetBackground, authMiddleware)
	api.GET("/background/choices", uc.GetUserBackgroundChoices, authMiddleware)
	api.DELETE("/:user_id", uc.DeleteUser, authMiddleware)
}